refactor: split browser formatting helpers
